# backend/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Six Seven API",
    "version": "1.0.0",
    "description": "Backend API for Six Seven - Talk the Alpha, Walk the Alpha",
    "contact": {
      "name": "Six Seven Support",
      "email": "support@sixseven.app"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/apple/callback": {
      "post": {
        "summary": "Apple OAuth callback",
        "description": "Handle Apple OAuth callback for authentication",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "Authorization code from Apple"
                  },
                  "id_token": {
                    "type": "string",
                    "description": "ID token from Apple"
                  },
                  "user": {
                    "type": "string",
                    "description": "User data from Apple (JSON string, first sign-in only)"
                  },
                  "error": {
                    "type": "string",
                    "description": "Error from Apple OAuth"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Redirect to app with authentication token"
          },
          "400": {
            "description": "No ID token received or email required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Apple OAuth callback (GET)",
        "description": "Handle Apple OAuth callback via GET (not recommended)",
        "responses": {
          "405": {
            "description": "Use POST method for Apple callback",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/apple/initiate": {
      "get": {
        "summary": "Initiate Apple OAuth",
        "description": "Redirect to Apple OAuth consent screen for authentication",
        "responses": {
          "302": {
            "description": "Redirect to Apple OAuth consent screen"
          },
          "500": {
            "description": "Failed to initiate Apple OAuth",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/check": {
      "post": {
        "summary": "Check user existence",
        "description": "Check if a user exists by email and their onboarding status",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User email address"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User check completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": {
                      "type": "boolean",
                      "description": "Whether the user exists"
                    },
                    "onboarded": {
                      "type": "boolean",
                      "description": "Whether the user has completed onboarding"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/email/send": {
      "post": {
        "summary": "Send verification email",
        "description": "Send a verification code to the user's email address",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address to send verification code to"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Failed to send email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/google/callback": {
      "get": {
        "summary": "Google OAuth callback (web)",
        "description": "Handle Google OAuth callback for web authentication",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            },
            "description": "Authorization code from Google"
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            },
            "description": "State parameter for redirect URI"
          },
          {
            "in": "query",
            "name": "error",
            "schema": {
              "type": "string"
            },
            "description": "Error from Google OAuth"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to app with authentication token"
          },
          "400": {
            "description": "Missing authorization code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Google Sign-In (native)",
        "description": "Authenticate with Google ID token for native apps",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idToken"
                ],
                "properties": {
                  "idToken": {
                    "type": "string",
                    "description": "Google ID token"
                  },
                  "accessToken": {
                    "type": "string",
                    "description": "Google access token (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT authentication token"
                    },
                    "onboarded": {
                      "type": "boolean",
                      "description": "Whether the user has completed onboarding"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "name": {
                          "type": "string",
                          "nullable": true
                        },
                        "picture": {
                          "type": "string",
                          "format": "uri",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing ID token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}